+++ index.js
@@ -1,5 +1,0 @@
-// Root project/index.js
-import { registerRootComponent } from 'expo';
-import App from './App';
-
-registerRootComponent(App);