+++ loginUI/index.js
@@ -1,4 +1,0 @@
-// loginUI/index.js
-import { registerRootComponent } from 'expo';
-import LoginNavigator from './src/screens/LoginNavigator';
-registerRootComponent(LoginNavigator);