+++ loginUI/src/screens/auth/login/LoginScreen.tsx
@@ -1,3 +1,4 @@
+// loginUI/login/LoginScreen
 import { AuthStackParamList } from '@login/navigation/types';
 import { useNavigation } from '@react-navigation/native';
 import { NativeStackNavigationProp } from '@react-navigation/native-stack';