+++ mainServiceUI/App.tsx
@@ -1,7 +1,0 @@
-// App.tsx
-import React from 'react';
-import MainServiceScreen from './src/screens/MainServiceScreen';
-
-export default function App() {
-  return <MainServiceScreen />;
-}